move: RedisInitialization이 테스트 뿐만아니라 스프링 처음 시작시에도 쓰여야 할 것 같아 위치 이동
